Przejście na drugą wersję aplikacji

+++ AlgorytmyDoTTP/AlgorytmyDoTTP/Struktura/Algorytmy/Ewolucyjny/Osobnik/OsobnikTSP.cs
@@ -27,7 +27,7 @@ namespace AlgorytmyDoTTP.Struktura.Algorytmy.Ewolucyjny.Osobnik
             throw new System.NotImplementedException();
         }
 
-        public override Dictionary<string, double[]> FunkcjaDopasowania(ReprezentacjaRozwiazania reprezentacjaGenotypu)
+        public override Dictionary<string, float[]> FunkcjaDopasowania(ReprezentacjaRozwiazania reprezentacjaGenotypu)
         {
             return problemOptymalizacyjny.ObliczZysk(Fenotyp(reprezentacjaGenotypu.ZwrocGenotyp1Wymiarowy()));
         }
