+++ AlgorytmyDoTTP/AlgorytmyDoTTP/Struktura/Algorytmy/Ewolucyjny/Osobnik/OsobnikTTP.cs
@@ -51,7 +51,7 @@ namespace AlgorytmyDoTTP.Struktura.Algorytmy.Ewolucyjny.Osobnik
             throw new NotImplementedException();
         }
 
-        public override Dictionary<string, double[]> FunkcjaDopasowania(ReprezentacjaRozwiazania reprezentacjaGenotypu)
+        public override Dictionary<string, float[]> FunkcjaDopasowania(ReprezentacjaRozwiazania reprezentacjaGenotypu)
         {
             return problemOptymalizacyjny.ObliczZysk(Fenotyp(reprezentacjaGenotypu.ZwrocGenotyp2Wymiarowy()));
         }
