+++ AlgorytmyDoTTP/AlgorytmyDoTTP/Struktura/Algorytmy/Wspinaczkowy/Wspinaczka/WspinaczkaKP.cs
@@ -13,7 +13,7 @@ namespace AlgorytmyDoTTP.Struktura.Algorytmy.Wspinaczkowy.Wspinaczka
     {
         public WspinaczkaKP(ALosowanie losowanie) : base(losowanie) {}
 
-        public override Dictionary<string, double[]> ZnajdzOptimum()
+        public override Dictionary<string, float[]> ZnajdzOptimum()
         {
             int poprawy = 0;
             ushort[] genotyp = reprezentacjaRozwiazania.ZwrocGenotyp1Wymiarowy(),
@@ -21,7 +21,7 @@ namespace AlgorytmyDoTTP.Struktura.Algorytmy.Wspinaczkowy.Wspinaczka
 
             AOsobnik osobnik = losowanie.ZwrocOsobnika();
             ProblemOptymalizacyjny problemOptymalizacyjny = osobnik.ZwrocInstancjeProblemu();
-            Dictionary<string, double[]> wynik = osobnik.FunkcjaDopasowania(reprezentacjaRozwiazania);
+            Dictionary<string, float[]> wynik = osobnik.FunkcjaDopasowania(reprezentacjaRozwiazania);
             ReprezentacjaRozwiazania tmpReprezentacjaRozwiazanie = new ReprezentacjaRozwiazania(tmpGenotyp);
 
             do
@@ -37,7 +37,7 @@ namespace AlgorytmyDoTTP.Struktura.Algorytmy.Wspinaczkowy.Wspinaczka
                     tmpGenotyp[i] = (ushort)((tmpGenotyp[i] == 0) ? 1 : 0);
                     tmpReprezentacjaRozwiazanie.ZmienGenotyp(tmpGenotyp);
 
-                    Dictionary<string, double[]> tmpWynik = osobnik.FunkcjaDopasowania(tmpReprezentacjaRozwiazanie);
+                    Dictionary<string, float[]> tmpWynik = osobnik.FunkcjaDopasowania(tmpReprezentacjaRozwiazanie);
 
                     if (tmpWynik["max"][0] > wynik["max"][0])
                     {
