+++ AlgorytmyDoTTP/AlgorytmyDoTTP/Struktura/ProblemyOptymalizacyjne/TTP/Model/ITTP.cs
@@ -5,6 +5,6 @@ namespace AlgorytmyDoTTP.Struktura.ProblemyOptymalizacyjne.TTP.Model
 {
     interface ITTP 
     {
-        Dictionary<string, double[]> ObliczWartoscFunkcjiCelu(double[] dlugosciTrasy, ushort[][] mapaPrzedmiotow, double[] ograniczeniaProblemu, ProblemPlecakowy problemPlecakowy);
+        Dictionary<string, float[]> ObliczWartoscFunkcjiCelu(float[] dlugosciTrasy, ushort[][] mapaPrzedmiotow, float[] ograniczeniaProblemu, ProblemPlecakowy problemPlecakowy);
     }
 }
